--- game.go
@@ -1,4 +1,0 @@
-package telegraph
-
-// CallbackGame A placeholder, currently holds no information. Use BotFather to set up your game.
-type CallbackGame struct{}
