+++ update_test.go
@@ -1,16 +1,58 @@
 package telegraph_test
 
 import (
-	"fmt"
 	"net/http"
-	"telegraph"
 	"testing"
+
+	"telegraph"
+
+	"fmt"
 
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/h2non/gock.v1"
 )
 
-func TestGetUpdateSuccess(t *testing.T) {
+func TestWebHookParseRequest_Success(t *testing.T) {
+	payload := []byte(`{
+		"update_id": 651868729,
+		"message": {
+			"message_id": 19,
+			"from": {
+				"id": 234234,
+				"is_bot": false,
+				"first_name": "Dimas",
+				"last_name": "Ragil T",
+				"username": "dynastymasra",
+				"language_code": "en-US"
+			},
+			"chat": {
+				"id": 23423423,
+				"first_name": "Dimas",
+				"last_name": "Ragil T",
+				"username": "dynastymasra",
+				"type": "private"
+			},
+			"date": 1508298329,
+			"text": "test text"
+		}
+	}`)
+
+	message, err := telegraph.WebHookParseRequest(payload)
+
+	assert.NotNil(t, message)
+	assert.NoError(t, err)
+}
+
+func TestWebHookParseRequest_Failed(t *testing.T) {
+	payload := []byte(`<-`)
+
+	message, err := telegraph.WebHookParseRequest(payload)
+
+	assert.Nil(t, message)
+	assert.Error(t, err)
+}
+
+func TestGetUpdates_Success(t *testing.T) {
 	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON(`{
 	  "ok": true,
 	  "result": [
@@ -41,193 +83,28 @@ func TestGetUpdateSuccess(t *testing.T) {
 
 	client := telegraph.NewClient("token")
 
-	model, res, err := client.GetUpdate().Commit()
+	model, res, err := client.GetUpdates().SetOffset(5).SetLimit(5).SetTimeout(5).SetAllowedUpdates("1", "2").Commit()
 
 	assert.NotNil(t, model)
 	assert.Equal(t, http.StatusOK, res.StatusCode)
 	assert.NoError(t, err)
 }
 
-func TestGetUpdateSetOffset(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON(`{
-	  "ok": true,
-	  "result": [
-		{
-		  "update_id": 1234567890,
-		  "message": {
-			"message_id": 238,
-			"from": {
-			  "id": 1234567890,
-			  "is_bot": true,
-			  "first_name": "Cubesoft",
-			  "username": "CubesoftBot"
-			},
-			"chat": {
-			  "id": 1234567890,
-			  "first_name": "cube",
-			  "last_name": "soft",
-			  "username": "cubesoft",
-			  "type": "private"
-			},
-			"date": 1510033702,
-			"text": "test"
-		  }
-		}
-	  ]
-	}`)
+func TestGetUpdates_Error(t *testing.T) {
+	gock.New(telegraph.BaseURL).Head(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusInternalServerError).JSON("")
 	defer gock.Off()
 
 	client := telegraph.NewClient("token")
 
-	model, res, err := client.GetUpdate().SetOffset(5).Commit()
-
-	assert.NotNil(t, model)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.NoError(t, err)
-}
-
-func TestGetUpdateSetLimit(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON(`{
-	  "ok": true,
-	  "result": [
-		{
-		  "update_id": 1234567890,
-		  "message": {
-			"message_id": 238,
-			"from": {
-			  "id": 1234567890,
-			  "is_bot": true,
-			  "first_name": "Cubesoft",
-			  "username": "CubesoftBot"
-			},
-			"chat": {
-			  "id": 1234567890,
-			  "first_name": "cube",
-			  "last_name": "soft",
-			  "username": "cubesoft",
-			  "type": "private"
-			},
-			"date": 1510033702,
-			"text": "test"
-		  }
-		}
-	  ]
-	}`)
-	defer gock.Off()
-
-	client := telegraph.NewClient("token")
-
-	model, res, err := client.GetUpdate().SetLimit(5).Commit()
-
-	assert.NotNil(t, model)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.NoError(t, err)
-}
-
-func TestGetUpdateSetTimeout(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON(`{
-	  "ok": true,
-	  "result": [
-		{
-		  "update_id": 1234567890,
-		  "message": {
-			"message_id": 238,
-			"from": {
-			  "id": 1234567890,
-			  "is_bot": true,
-			  "first_name": "Cubesoft",
-			  "username": "CubesoftBot"
-			},
-			"chat": {
-			  "id": 1234567890,
-			  "first_name": "cube",
-			  "last_name": "soft",
-			  "username": "cubesoft",
-			  "type": "private"
-			},
-			"date": 1510033702,
-			"text": "test"
-		  }
-		}
-	  ]
-	}`)
-	defer gock.Off()
-
-	client := telegraph.NewClient("token")
-
-	model, res, err := client.GetUpdate().SetTimeout(5).Commit()
-
-	assert.NotNil(t, model)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.NoError(t, err)
-}
-
-func TestGetUpdateSetAllowUpdate(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON(`{
-	  "ok": true,
-	  "result": [
-		{
-		  "update_id": 1234567890,
-		  "message": {
-			"message_id": 238,
-			"from": {
-			  "id": 1234567890,
-			  "is_bot": true,
-			  "first_name": "Cubesoft",
-			  "username": "CubesoftBot"
-			},
-			"chat": {
-			  "id": 1234567890,
-			  "first_name": "cube",
-			  "last_name": "soft",
-			  "username": "cubesoft",
-			  "type": "private"
-			},
-			"date": 1510033702,
-			"text": "test"
-		  }
-		}
-	  ]
-	}`)
-	defer gock.Off()
-
-	client := telegraph.NewClient("token")
-
-	model, res, err := client.GetUpdate().SetAllowedUpdates("1", "2", "3").Commit()
-
-	assert.NotNil(t, model)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.NoError(t, err)
-}
-
-func TestGetUpdateError(t *testing.T) {
-	gock.New(telegraph.BaseURL).Head(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusOK).JSON("")
-	defer gock.Off()
-
-	client := telegraph.NewClient("token")
-
-	model, res, err := client.GetUpdate().Commit()
+	model, res, err := client.GetUpdates().SetOffset(5).SetLimit(5).SetTimeout(5).SetAllowedUpdates("1", "2").Commit()
 
 	assert.Nil(t, model)
 	assert.Equal(t, http.StatusInternalServerError, res.StatusCode)
 	assert.Error(t, err)
 }
 
-func TestGetUpdateFailedUnmarshal(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusBadGateway).XML("")
-	defer gock.Off()
-
-	client := telegraph.NewClient("token")
-
-	model, res, err := client.GetUpdate().Commit()
-
-	assert.Nil(t, model)
-	assert.Equal(t, http.StatusBadGateway, res.StatusCode)
-	assert.Error(t, err)
-}
-
-func TestGetUpdateFailed(t *testing.T) {
-	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusUnauthorized).XML(`{
+func TestGetUpdates_Failed(t *testing.T) {
+	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetUpdate, "token")).Reply(http.StatusUnauthorized).JSON(`{
 		"ok": false,
 		"error_code": 401,
 		"description": "Unauthorized"
@@ -236,7 +113,7 @@ func TestGetUpdateFailed(t *testing.T) {
 
 	client := telegraph.NewClient("token")
 
-	model, res, err := client.GetUpdate().Commit()
+	model, res, err := client.GetUpdates().Commit()
 
 	assert.Nil(t, model)
 	assert.Equal(t, http.StatusUnauthorized, res.StatusCode)
