+++ webhook.go
@@ -1,14 +1,59 @@
 package telegraph
 
-import "encoding/json"
+import (
+	"fmt"
 
-// WebHookParseRequest function for parse request from telegram web hook, return struct Update if success
-func WebHookParseRequest(r []byte) (*Update, error) {
-	update := &Update{}
+	"net/http"
 
-	if err := json.Unmarshal(r, update); err != nil {
-		return nil, err
+	"github.com/cenkalti/backoff"
+	"github.com/parnurzeal/gorequest"
+)
+
+type (
+	// WebHookInfoResponse struct to handle request and response telegram api
+	WebHookInfoResponse struct {
+		Client  *Client
+		Request *gorequest.SuperAgent
+	}
+)
+
+// GetWebHookInfo Use this method to get current webhook status. Requires no parameters.
+// On success, returns a WebhookInfo object. If the bot is using getUpdates,
+// will return an object with the url field empty.
+func (client *Client) GetWebHookInfo() *WebHookInfoResponse {
+	url := client.baseURL + fmt.Sprintf(EndpointGetWebHookInfo, client.accessToken)
+	request := gorequest.New().Get(url).Set(UserAgentHeader, UserAgent+"/"+Version)
+
+	return &WebHookInfoResponse{
+		Client:  client,
+		Request: request,
+	}
+}
+
+// Commit execute request to telegram
+func (info *WebHookInfoResponse) Commit() (*WebhookInfo, *http.Response, error) {
+	var errs []error
+	var body []byte
+	res := &http.Response{}
+	model := struct {
+		ErrorResponse
+		Result *WebhookInfo `json:"result,omitempty"`
+	}{}
+
+	operation := func() error {
+		res, body, errs = info.Request.EndStruct(&model)
+		if len(errs) > 0 {
+			return errs[0]
+		}
+		return nil
 	}
 
-	return update, nil
+	if err := backoff.Retry(operation, info.Client.expBackOff); err != nil {
+		return nil, MakeHTTPResponse(info.Request), err
+	}
+	if res.StatusCode != http.StatusOK {
+		return nil, res, fmt.Errorf("%v %v", model.ErrorCode, model.Description)
+	}
+
+	return model.Result, res, nil
 }
