+++ webhook_test.go
@@ -1,48 +1,62 @@
 package telegraph_test
 
 import (
+	"fmt"
+	"net/http"
 	"telegraph"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"gopkg.in/h2non/gock.v1"
 )
 
-func TestWebHookParseRequestSuccess(t *testing.T) {
-	payload := []byte(`{
-		"update_id": 651868729,
-		"message": {
-			"message_id": 19,
-			"from": {
-				"id": 234234,
-				"is_bot": false,
-				"first_name": "Dimas",
-				"last_name": "Ragil T",
-				"username": "dynastymasra",
-				"language_code": "en-US"
-			},
-			"chat": {
-				"id": 23423423,
-				"first_name": "Dimas",
-				"last_name": "Ragil T",
-				"username": "dynastymasra",
-				"type": "private"
-			},
-			"date": 1508298329,
-			"text": "test text"
+func TestGetWebHookInfo_Success(t *testing.T) {
+	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetWebHookInfo, "token")).Reply(http.StatusOK).JSON(`{
+		"ok": true,
+		"result": {
+			"url": "https://www.cube.com/webhook",
+			"has_custom_certificate": false,
+			"pending_update_count": 0,
+			"max_connections": 100
 		}
 	}`)
+	defer gock.Off()
 
-	message, err := telegraph.WebHookParseRequest(payload)
+	client := telegraph.NewClient("token")
 
-	assert.NotNil(t, message)
+	info, res, err := client.GetWebHookInfo().Commit()
+
+	assert.NotNil(t, info)
+	assert.Equal(t, http.StatusOK, res.StatusCode)
 	assert.NoError(t, err)
 }
 
-func TestWebHookParseRequestFailed(t *testing.T) {
-	payload := []byte(`<-`)
+func TestGetWebHookInfo_Error(t *testing.T) {
+	gock.New(telegraph.BaseURL).Head(fmt.Sprintf(telegraph.EndpointGetWebHookInfo, "token")).Reply(http.StatusInternalServerError).JSON("")
+	defer gock.Off()
 
-	message, err := telegraph.WebHookParseRequest(payload)
+	client := telegraph.NewClient("token")
 
-	assert.Nil(t, message)
+	info, res, err := client.GetWebHookInfo().Commit()
+
+	assert.Nil(t, info)
+	assert.Equal(t, http.StatusInternalServerError, res.StatusCode)
+	assert.Error(t, err)
+}
+
+func TestGetWebHookInfo_Failed(t *testing.T) {
+	gock.New(telegraph.BaseURL).Get(fmt.Sprintf(telegraph.EndpointGetWebHookInfo, "token")).Reply(http.StatusNotFound).JSON(`{
+		"ok": false,
+		"error_code": 404,
+		"description": "Not Found: method not found"
+	}`)
+	defer gock.Off()
+
+	client := telegraph.NewClient("token")
+
+	info, res, err := client.GetWebHookInfo().Commit()
+
+	assert.Nil(t, info)
+	assert.Equal(t, http.StatusNotFound, res.StatusCode)
 	assert.Error(t, err)
 }
